Refactor admin routes into modular structure for better maintainability
Split monolithic admin routes.py (2,448 lines) into logical modules:
- routes/auth.py: Authentication & session management
- routes/dashboard.py: Main dashboard with stats
- routes/users.py: User management operations
- routes/apps.py: App lifecycle management
- routes/questions.py: Question management system
- routes/llm.py: LLM configuration & analytics

Benefits:
- Improved code organization and maintainability
- Reduced memory footprint per module
- Easier debugging and testing
- Better separation of concerns

Also fixed parse_recipe_metadata bug in content routes.
